backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.7
decorator==4.4.2
Django==3.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.4.0
django-extensions==3.0.3
django-filter==2.3.0
django-mysql==3.8.1
django-phonenumber-field==5.0.0
django-rest-registration==0.5.6
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
drf-yasg==1.17.1
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.11
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.3.0
